Work in progress - use builder to create all service iterator instances
Builders are used to ensure that filters are applied to the service
request where possible, with a fallback to filtering data returned from
the service provider.

Something like this was done with PhotoIterator, but the new code
applies to all iterators (not just photo iterators), and is easier to
use.

Still to do: roll builder out to all iterators (implemented for Flickr)
and then remove PhotoIterator and PhotoIteratorBuilder