Add Windows registry user store without extra slashes
java.util.prefs.Preferences adds a slash before upper case letters in
keys and values.

This adds an alternative user store for Windows only which does not use
the leading slashes.

Requires JNA. If JNA is not present, or not Windows, will fallback to
the previous store for user tokens and secrets.